Server: Use scylla CachingSession
Using `.query_*` methods on `Session` object for queries with parameters
requires 2 round trips to server: first the request is prepared, and
only then it is executed.

This is because serializing arguments for a request requires knowing the
types of those arguments - which are only known in prepared statement.

For more information on that, refer to the following chapter of
Scylla Rust Driver book:
https://rust-driver.docs.scylladb.com/stable/queries/queries.html

There are 2 possible fixes:
- Using `PreparedStatement` and `.execute_*` methods.
- Using `CachingSession` instead of `Session` - it handles preparation
  internally and caches prepared statements.

The second solution is simpler to implement as it doesn't require
many changes in the code. It will also be more resilient as future
changes won't accidentally re-introduce the problem by issuing
unprepared statements.

I've decided to not use `CachingSession` in call to `initialize_schema`.
Queries issued there don't have parameters, so they don't suffer from
the double round trip problem. Also, those are schema changing queries,
so they don't benefit from being prepared.